// generated/api/models/bankAccountDto.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */
import type { DecimalValue } from './decimalValue';
import type { BankAccountType } from './bankAccountType';
import type { BankAccountStatus } from './bankAccountStatus';
import type { Timestamp } from './timestamp';

export interface BankAccountDto {
  /** @nullable */
  bankAccountId?: string | null;
  /** @nullable */
  agreementId?: string | null;
  /** @nullable */
  clientId?: string | null;
  balance?: DecimalValue;
  type?: BankAccountType;
  status?: BankAccountStatus;
  createdAt?: Timestamp;
  updatedAt?: Timestamp;
}